Main/requirements-security.txt
ultralytics>=8.3.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
opencv-python>=4.7.0
Pillow>=9.5.0
numpy>=1.24.0
soundfile>=0.12.0
sounddevice>=0.4.6
pygame>=2.5.0

